Impl session persistence with GPG instead of token
The old system left the 1Password session token exposed in a cleartext
file, which could then be stolen during the session timeout. This has
been replaced by a 1p-specific GPG key, which is then used for
encrypting the persistent configuration (which now includes the master
password, but excludes the session token). Session now persists as much
as the keyring remains unlocked, and is only accessible to the keyring
owner.